+++ engine/lite/k8sDeploymentManager/src/main/scala/pl/touk/nussknacker/k8s/manager/K8sProcessStateDefinitionManager.scala
@@ -1,18 +1,8 @@
 package pl.touk.nussknacker.k8s.manager
 
-import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
-import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager}
-import pl.touk.nussknacker.engine.api.deployment.{OverridingProcessStateDefinitionManager, ProcessState}
-import skuber.apps.v1.Deployment
+import pl.touk.nussknacker.engine.api.deployment.OverridingProcessStateDefinitionManager
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleProcessStateDefinitionManager
 
 object K8sProcessStateDefinitionManager extends OverridingProcessStateDefinitionManager(
   delegate = SimpleProcessStateDefinitionManager
-) {
-
-  def errorMultipleJobsRunning(duplicates: List[Deployment]): ProcessState = {
-    processState(ProblemStateStatus.multipleJobsRunning)
-      .copy(
-        errors = List(s"Expected one deployment, instead: ${duplicates.map(_.metadata.name).mkString(", ")}")
-      )
-  }
-}
+)